third_party/blink/web_tests/fast/dom/StyleSheet/detached-stylesheet-without-wrapper-expected.txt

Test that detached CSSOM subtrees keeps a reference to CSS stylesheets lacking wrappers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rule.parentStyleSheet is not null
PASS successfullyParsed is true

TEST COMPLETE
